better fold column implementation
Sets foldcolumn to +1, which translates to &textwidth+1. This
automatically adjusts based on &textwidth, and disables the colorcolumn
when &textwidth is not set. So we can remove all the stupid cc settings
in all the autocmd sections.

Also, makes cc only one char wide which looks better on a wide layout
with background transparency.